Move datetime readers into a separate module
